Added deep find for professor
Added deep_get_professor_by_school_and_name(college: School, professor_name: str):

It is the same as get_professor_by_school_and_name, but splits the name, creating a list of possible names. Then it searches for the professor with the highest similarity rating when compared to the passed professor name.